Update the demo to use MediaSourceHandle+srcObject
With [spec PR #306](https://github.com/w3c/media-source/pull/306) merged
and with implementation landed corresponding to that spec already landed
in Chromium as of 105.0.5180, MSE-in-Workers requires a
MediaSourceHandle transferred from a dedicated worker to attach a worker
MediaSource to a media element. This change updates the demo to check
for the necessary support and, use handle and srcObject for worker
mediasource attachment (still leaving legacy MediaSource object URL for
main thread mediasource attachments), and updates various instructions
and READMEs accordingly.